+++ clean_python/s3/s3_provider.py
@@ -7,8 +7,9 @@ try:
     import aioboto3
 except ImportError:
     aioboto3 = None
-
 from botocore.client import Config
+
+from clean_python import Provider
 
 from .s3_bucket_options import S3BucketOptions
 
@@ -20,19 +21,14 @@ __all__ = ["S3BucketOptions", "S3BucketProvider"]
 logger = logging.getLogger(__name__)
 
 
-class S3BucketProvider:
+class S3BucketProvider(Provider):
     def __init__(self, options: S3BucketOptions):
         self.options = options
 
-    @property
-    def bucket(self) -> str:
-        return self.options.bucket
-
-    @property
-    def client(self) -> "S3Client":
+    async def connect(self) -> None:
         assert aioboto3 is not None
-        session = aioboto3.Session()
-        return session.client(
+        self._session = aioboto3.Session()
+        self._client = await self._session.client(
             "s3",
             endpoint_url=self.options.url,
             aws_access_key_id=self.options.access_key,
@@ -47,4 +43,15 @@ class S3BucketProvider:
                 },
             ),
             use_ssl=self.options.url.startswith("https"),
-        )
+        ).__aenter__()
+
+    async def disconnect(self) -> None:
+        await self._client.close()
+
+    @property
+    def bucket(self) -> str:
+        return self.options.bucket
+
+    @property
+    def client(self) -> "S3Client":
+        return self._client