+++ integration_tests/test_int_api_provider.py
@@ -13,10 +13,18 @@ async def fake_token():
 
 
 @pytest.fixture
-def provider(fastapi_example_app) -> ApiProvider:
+def tenant_context() -> Tenant:
     ctx.tenant = Tenant(id=2, name="")
-    yield ApiProvider(fastapi_example_app + "/", fake_token)
+    yield ctx.tenant
     ctx.tenant = None
+
+
+@pytest.fixture
+async def provider(fastapi_example_app, tenant_context) -> ApiProvider:
+    provider = ApiProvider(fastapi_example_app + "/", fake_token)
+    await provider.connect()
+    yield provider
+    await provider.disconnect()
 
 
 async def test_request_params(provider: ApiProvider):