+++ tests/oauth2/test_client_credentials.py
@@ -31,15 +31,25 @@ def settings() -> OAuth2CCSettings:
 
 
 @pytest.fixture
-def gateway(settings) -> CCTokenGateway:
-    with mock.patch(MODULE + ".ApiProvider", autospec=True):
-        yield CCTokenGateway(settings)
+def provider_m():
+    with mock.patch(MODULE + ".ApiProvider", autospec=True) as x:
+        yield x.return_value
 
 
 @pytest.fixture
-def sync_gateway(settings) -> SyncCCTokenGateway:
-    with mock.patch(MODULE + ".SyncApiProvider", autospec=True):
-        yield SyncCCTokenGateway(settings)
+def gateway(settings, provider_m) -> CCTokenGateway:
+    return CCTokenGateway(settings)
+
+
+@pytest.fixture
+def sync_provider_m():
+    with mock.patch(MODULE + ".SyncApiProvider", autospec=True) as x:
+        yield x.return_value
+
+
+@pytest.fixture
+def sync_gateway(settings, sync_provider_m) -> SyncCCTokenGateway:
+    return SyncCCTokenGateway(settings)
 
 
 @pytest.mark.parametrize(
@@ -55,91 +65,95 @@ def test_is_token_usable(expires_in, leeway, expected):
     assert is_token_usable(token, leeway) is expected
 
 
-async def test_fetch_token(gateway: CCTokenGateway):
-    gateway.provider.request.return_value = {"access_token": "foo"}
+async def test_fetch_token(gateway: CCTokenGateway, provider_m):
+    provider_m.request.return_value = {"access_token": "foo"}
 
     token = await gateway._fetch_token()
 
     assert token == "foo"
 
-    gateway.provider.request.assert_awaited_once_with(
+    provider_m.connect.assert_awaited_once()
+    provider_m.request.assert_awaited_once_with(
         method="POST",
         path="",
         fields={"grant_type": "client_credentials", "scope": "all"},
         timeout=1.0,
     )
+    provider_m.disconnect.assert_awaited_once()
 
 
-async def test_fetch_token_cache(gateway: CCTokenGateway):
+async def test_fetch_token_cache(gateway: CCTokenGateway, provider_m):
     # empty cache: provider gets called
     token = get_token({})
-    gateway.provider.request.return_value = {"access_token": token}
+    provider_m.request.return_value = {"access_token": token}
     actual = await gateway.fetch_token()
     assert actual == token
-    assert gateway.provider.request.called
+    assert provider_m.request.called
 
-    gateway.provider.request.reset_mock()
+    provider_m.request.reset_mock()
 
     # cache is filled: provider is not called
     actual = await gateway.fetch_token()
     assert actual == token
-    assert not gateway.provider.request.called
+    assert not provider_m.request.called
 
-    gateway.provider.request.reset_mock()
+    provider_m.request.reset_mock()
 
     # token is not usable so it is refreshed:
     with mock.patch(MODULE + ".is_token_usable", side_effect=(False, True)):
         actual = await gateway.fetch_token()
         assert actual == token
-        assert gateway.provider.request.called
+        assert provider_m.request.called
 
 
-def test_fetch_token_sync(sync_gateway: SyncCCTokenGateway):
-    sync_gateway.provider.request.return_value = {"access_token": "foo"}
+def test_fetch_token_sync(sync_gateway: SyncCCTokenGateway, sync_provider_m):
+    sync_provider_m.request.return_value = {"access_token": "foo"}
 
     token = sync_gateway._fetch_token()
 
     assert token == "foo"
 
-    sync_gateway.provider.request.assert_called_once_with(
+    sync_provider_m.connect.assert_called_once()
+    sync_provider_m.request.assert_called_once_with(
         method="POST",
         path="",
         fields={"grant_type": "client_credentials", "scope": "all"},
         timeout=1.0,
     )
+    sync_provider_m.disconnect.assert_called_once()
 
 
-def test_fetch_token_sync_cache(sync_gateway: SyncCCTokenGateway):
+def test_fetch_token_sync_cache(sync_gateway: SyncCCTokenGateway, sync_provider_m):
     # empty cache: provider gets called
     token = get_token({})
-    sync_gateway.provider.request.return_value = {"access_token": token}
+    sync_provider_m.request.return_value = {"access_token": token}
     actual = sync_gateway.fetch_token()
     assert actual == token
-    assert sync_gateway.provider.request.called
+    assert sync_provider_m.request.called
 
-    sync_gateway.provider.request.reset_mock()
+    sync_provider_m.request.reset_mock()
 
     # cache is filled: provider is not called
     actual = sync_gateway.fetch_token()
     assert actual == token
-    assert not sync_gateway.provider.request.called
+    assert not sync_provider_m.request.called
 
-    sync_gateway.provider.request.reset_mock()
+    sync_provider_m.request.reset_mock()
 
     # token is not usable so it is refreshed:
     with mock.patch(MODULE + ".is_token_usable", side_effect=(False, True)):
         actual = sync_gateway.fetch_token()
         assert actual == token
-        assert sync_gateway.provider.request.called
+        assert sync_provider_m.request.called
 
 
-async def test_fetch_headers(gateway: CCTokenGateway):
-    gateway.provider.request.return_value = {"access_token": "foo"}
+async def test_fetch_headers(gateway: CCTokenGateway, provider_m):
+    provider_m.request.return_value = {"access_token": "foo"}
 
     await gateway.fetch_headers() == {"Authorization": "Bearer foo"}
 
 
-def test_fetch_headers_sync(sync_gateway: SyncCCTokenGateway):
-    sync_gateway.provider.request.return_value = {"access_token": "foo"}
+def test_fetch_headers_sync(sync_gateway: SyncCCTokenGateway, sync_provider_m):
+    sync_provider_m.request.return_value = {"access_token": "foo"}
 
     assert sync_gateway.fetch_headers() == {"Authorization": "Bearer foo"}